create tools.py for update_server_configuration func
